refactor[#38]: rename logic package to service
