+++ src/main/java/com/codecool/codecoin/configuration/Configuration.java
@@ -1,9 +1,8 @@
 package com.codecool.codecoin.configuration;
 
-import com.codecool.codecoin.logic.Calculator;
+import com.codecool.codecoin.service.Calculator;
 import com.codecool.codecoin.model.Portfolio;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Scope;
 import org.springframework.web.reactive.function.client.WebClient;
 
 /**
